Extract stub initializer
