+++ gemmforge/generators/__init__.py
@@ -1,4 +1,4 @@
-from .csa_generator import CsaGenerator
 from .abstract_generator import AbstractGenerator
+from .csa_generator import CsaGenerator
 from .gemm_generator import GemmGenerator
 from .abstract_gemmlike_generator import GemmLikeGenerator
